requirements.txt
certifi==2020.6.20
chardet==3.0.4
idna==2.10
peewee==3.13.3
PyMySQL==0.10.0
requests==2.24.0
urllib3==1.25.10
mysql-connector-python==8.0.22


